feat: Extracted state/session and UI notification handlers into focused…
- src/webview/handlers/state-handlers.ts
- src/webview/handlers/ui-notification-handlers.ts
- src/webview/message-handler.ts
- src/webview/__tests__/accessibility.test.ts

GSD-Task: S02/T02
